Refactoring tool class NetUtils (#2316)
* Refactoring tool class NetUtils

* Remove redundant code

* Remove redundant code

* Remove redundant code

* Remove redundant code

* Remove redundant code

* import package
