+++ eventmesh-common/src/main/java/org/apache/eventmesh/common/Constants.java
@@ -155,6 +155,8 @@ public class Constants {
 
     public static final String EMPTY = "";
 
+    public static final int SUCCESS_CODE = 200;
+
     /**
      * GRPC PROTOCOL
      */
